Fixed impulse component
